requirements.txt
matplotlib
numpy
pandas
scikit-learn
tensorflow-gpu==2.1.0
opencv-contrib-python